@@ -9,7 +9,7 @@ global using Microsoft.eShopOnContainers.Services.Catalog.API.IntegrationEvents;
 global using Grpc.Core;
 global using Microsoft.AspNetCore.Hosting;
 global using Microsoft.AspNetCore.Http;
-global using Microsoft.AspNetCore.Builder;
+global using Microsoft.AspNetCore.Builder;
 global using Microsoft.AspNetCore.Mvc.Filters;
 global using Microsoft.AspNetCore.Mvc;
 global using Microsoft.AspNetCore.Server.Kestrel.Core;
@@ -52,7 +52,6 @@ global using System;
 global using Microsoft.eShopOnContainers.Services.Catalog.API.Infrastructure.Filters;
 global using HealthChecks.UI.Client;
 global using Microsoft.AspNetCore.Diagnostics.HealthChecks;
-global using Azure.Messaging.ServiceBus;
 global using Microsoft.eShopOnContainers.BuildingBlocks.EventBus;
 global using Microsoft.eShopOnContainers.BuildingBlocks.EventBusRabbitMQ;
 global using Microsoft.eShopOnContainers.BuildingBlocks.EventBusServiceBus;
@@ -61,3 +60,4 @@ global using Microsoft.Extensions.Diagnostics.HealthChecks;
 global using Microsoft.OpenApi.Models;
 global using RabbitMQ.Client;
 global using System.Reflection;
+global using Microsoft.Extensions.FileProviders;
