@@ -1,19 +1,15 @@
-global using Microsoft.eShopOnContainers.Services.Identity.API.Extensions;
-global using System.IO.Compression;
-global using Autofac.Extensions.DependencyInjection;
-global using Autofac;
 global using Azure.Core;
 global using Azure.Identity;
 global using HealthChecks.UI.Client;
 global using IdentityModel;
-global using IdentityServer4.EntityFramework.DbContexts;
-global using IdentityServer4.EntityFramework.Mappers;
-global using IdentityServer4.EntityFramework.Options;
-global using IdentityServer4.Models;
-global using IdentityServer4.Services;
-global using IdentityServer4.Stores;
-global using IdentityServer4.Validation;
-global using IdentityServer4;
+global using Duende.IdentityServer;
+global using Duende.IdentityServer.Configuration;
+global using Duende.IdentityServer.Events;
+global using Duende.IdentityServer.Extensions;
+global using Duende.IdentityServer.Models;
+global using Duende.IdentityServer.Services;
+global using Duende.IdentityServer.Stores;
+global using Duende.IdentityServer.Validation;
 global using Microsoft.AspNetCore.Authentication;
 global using Microsoft.AspNetCore.Authorization;
 global using Microsoft.AspNetCore.Builder;
@@ -23,20 +19,17 @@ global using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 global using Microsoft.AspNetCore.Identity;
 global using Microsoft.AspNetCore.Mvc.Rendering;
 global using Microsoft.AspNetCore.Mvc;
-global using Microsoft.AspNetCore;
-global using Microsoft.EntityFrameworkCore.Design;
+global using Microsoft.AspNetCore.Mvc.Filters;
 global using Microsoft.EntityFrameworkCore.Infrastructure;
 global using Microsoft.EntityFrameworkCore.Metadata;
 global using Microsoft.EntityFrameworkCore.Migrations;
 global using Microsoft.EntityFrameworkCore;
-global using Microsoft.eShopOnContainers.Services.Identity.API.Certificates;
-global using Microsoft.eShopOnContainers.Services.Identity.API.Configuration;
+
+global using Microsoft.eShopOnContainers.Services.Identity.API;
 global using Microsoft.eShopOnContainers.Services.Identity.API.Data;
-global using Microsoft.eShopOnContainers.Services.Identity.API.Devspaces;
-global using Microsoft.eShopOnContainers.Services.Identity.API.Models.AccountViewModels;
+global using Microsoft.eShopOnContainers.Services.Identity.API.Configuration;
 global using Microsoft.eShopOnContainers.Services.Identity.API.Models;
 global using Microsoft.eShopOnContainers.Services.Identity.API.Services;
-global using Microsoft.eShopOnContainers.Services.Identity.API;
 global using Microsoft.Extensions.Configuration;
 global using Microsoft.Extensions.DependencyInjection;
 global using Microsoft.Extensions.Diagnostics.HealthChecks;
@@ -44,20 +37,17 @@ global using Microsoft.Extensions.Hosting;
 global using Microsoft.Extensions.Logging;
 global using Microsoft.Extensions.Options;
 global using Polly;
-global using Serilog;
-global using StackExchange.Redis;
 global using System.Collections.Generic;
 global using System.ComponentModel.DataAnnotations;
 global using System.Data.SqlClient;
 global using System.IdentityModel.Tokens.Jwt;
 global using System.IO;
 global using System.Linq;
-global using System.Reflection;
 global using System.Security.Claims;
-global using System.Security.Cryptography.X509Certificates;
 global using System.Text.RegularExpressions;
 global using System.Threading.Tasks;
 global using System;
+global using Microsoft.AspNetCore.Http;
 
 
 
