@@ -1,4 +1,4 @@
-﻿@if (ViewContext.ModelState.IsValid == false)
+@if (ViewContext.ModelState.IsValid == false)
 {
     <div class="alert alert-danger">
         <strong>Error</strong>
