@@ -1,1 +1,2 @@
-﻿@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
+@using IdentityServerHost.Quickstart.UI
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
