@@ -36,8 +36,12 @@ namespace Ordering.BackgroundTasks.Services
                 _logger.LogDebug("GracePeriodManagerService background task is doing background work.");
 
                 CheckConfirmedGracePeriodOrders();
-
-                await Task.Delay(_settings.CheckUpdateTime, stoppingToken);
+                try {
+                    await Task.Delay(_settings.CheckUpdateTime, stoppingToken);
+                }
+                catch (TaskCanceledException exception) {
+                    _logger.LogCritical(exception, "TaskCanceledException Error", exception.Message);
+                }
             }
 
             _logger.LogDebug("GracePeriodManagerService background task is stopping.");
