@@ -1,7 +1,5 @@
-﻿global using Devspaces.Support;
-global using HealthChecks.UI.Client;
+﻿global using HealthChecks.UI.Client;
 global using Microsoft.AspNetCore.Authentication.Cookies;
-global using Microsoft.AspNetCore.Authentication.JwtBearer;
 global using Microsoft.AspNetCore.Authentication.OpenIdConnect;
 global using Microsoft.AspNetCore.Authentication;
 global using Microsoft.AspNetCore.Authorization;
