@@ -1,5 +1,5 @@
 import { Component, OnInit }    from '@angular/core';
-import { Observable, Subscription } from 'rxjs';
+import { Observable, Subscription, throwError } from 'rxjs';
 import { catchError }           from 'rxjs/operators';
 
 import { CatalogService }       from './catalog.service';
@@ -121,7 +121,7 @@ export class CatalogComponent implements OnInit {
 
     private handleError(error: any) {
         this.errorReceived = true;
-        return Observable.throw(error);
+        return throwError(() => error);
     }
 }
 
